moved parts if centerPlayer() to the utils class to make it available to other modules
