+++ src/main/java/me/zeroeightsix/kami/util/combat/SurroundUtils.kt
@@ -5,7 +5,10 @@ import net.minecraft.block.Block
 import net.minecraft.entity.Entity
 import net.minecraft.init.Blocks
 import net.minecraft.util.math.BlockPos
+import net.minecraft.util.math.MathHelper
+import net.minecraft.util.math.Vec3d
 import kotlin.math.floor
+import kotlin.math.round
 
 /**
  * @author Xiaro
@@ -57,6 +60,36 @@ object SurroundUtils {
     @JvmStatic
     fun checkBlock(block: Block): Boolean {
         return block == Blocks.BEDROCK || block == Blocks.OBSIDIAN || block == Blocks.ENDER_CHEST || block == Blocks.ANVIL
+    }
+
+    @JvmStatic
+    fun centerPlayer(tp: Boolean): Boolean {
+        val centerDiff = getCenterDiff()
+        val centered = isCentered()
+        if (!centered) {
+            mc.player.setVelocity(0.0, -5.0, 0.0)
+            if (tp) {
+                val posX = mc.player.posX + MathHelper.clamp(centerDiff.x, -0.2, 0.2)
+                val posZ = mc.player.posZ + MathHelper.clamp(centerDiff.z, -0.2, 0.2)
+                mc.player.setPosition(posX, mc.player.posY, posZ)
+            } else {
+                mc.player.motionX = MathHelper.clamp(centerDiff.x / 2.0, -0.2, 0.2)
+                mc.player.motionZ = MathHelper.clamp(centerDiff.z / 2.0, -0.2, 0.2)
+            }
+        }
+        return centered
+    }
+
+    private fun isCentered(): Boolean {
+        return getCenterDiff().length() < 0.2
+    }
+
+    private fun getCenterDiff(): Vec3d {
+        return Vec3d(roundToCenter(mc.player.posX), mc.player.posY, roundToCenter(mc.player.posZ)).subtract(mc.player.positionVector)
+    }
+
+    private fun roundToCenter(doubleIn: Double): Double {
+        return round(doubleIn + 0.5) - 0.5
     }
 
     enum class HoleType {
